solidstructuralsimulation class done
